fix(#684): block Azure__ tool names and '### Related tools' sections from skills output (#685)
* fix(#684): block Azure__ tool names and ### Related tools from skills output

- SkillPageValidator: add POLICY errors for 'Azure__' pattern and
  '### Related tools' heading — these are implementation details that
  must never appear in customer-facing articles (§4.2)
- AcrolinxPostProcessor: add StripRelatedToolsSection() and
  StripMcpToolNameLines() as steps 0a/0b in Process(), running before
  all other transforms as defence-in-depth
- Tests: 2 new validator tests + 7 new post-processor tests (445 total,
  all passing)
- CHANGELOG: document fix under [Unreleased]


* fix: correct TotalTokenCount type mismatch after SDK update

The TotalTokenCount property changed from long to int? in the updated
Azure.AI.Inference SDK. Pattern match now uses int instead of long.


---------

+++ skills-generation/SkillsGen.Core/Generation/AzureOpenAiRewriter.cs
@@ -214,7 +214,7 @@ public class AzureOpenAiRewriter : ILlmRewriter
                     UserPrompt = userPrompt,
                     ResponseContent = result,
                     Model = _modelName,
-                    TotalTokens = response.Value.Usage?.TotalTokenCount is long totalTokenCount ? (int?)totalTokenCount : null,
+                    TotalTokens = response.Value.Usage?.TotalTokenCount is int totalTokenCount ? (int?)totalTokenCount : null,
                     DurationMs = sw.ElapsedMilliseconds,
                     RetryCount = attempt
                 });
